Make it deployable to multiple hosts! Fix up dodgy nginx installation
New issue: use a less coupled pipeline
This pipeline would use only artifacts downloaded from an HTTP endpoint.
Thus we will not need to deal with installing SBT.
The artifact server can do all the triggering of deploys via an SSH command.
Then we only have to worry about how to do the deploy & blue-green swap in this SSH command.

**Bonus**: since it's push, is it possible to install the initial artifact and then push later artifacts without having to set up an HTTP server?
There's a huge benefit in not having to maintain an HTTP server.
If we scp it and then call a deploy command life could be easier.
But how do we deal with the initial download?
Or do we need one at all?

This sort of deployment might not work so well with Vagrant and local development though.

Also removing the watcher service as a result of this idea. It's not necessary to overload the node with too many concerns, & it's better to keep it flaaat.